% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_filter.R
\name{apply_value_filter}
\alias{apply_value_filter}
\title{Apply value filter conditions with validation}
\usage{
apply_value_filter(data, conditions, r_expr_validated, r_conditions_validated)
}
\arguments{
\item{data}{Input data frame}

\item{conditions}{List of filter conditions}

\item{r_expr_validated}{Reactive value for validated expression}

\item{r_conditions_validated}{Reactive value for validated conditions}
}
\description{
Apply value filter conditions with validation
}
