% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.dplyr-package}
\alias{blockr.dplyr}
\alias{blockr.dplyr-package}
\title{blockr.dplyr: Data wranglingling blocks for 'blockr.core'}
\description{
Provides access to 'dplyr'-powered data wrangling as blocks for 'blockr.core'.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

Authors:
\itemize{
  \item David Granjon \email{david@cynkra.com}
  \item Christoph Sax \email{christoph@cynkra.com}
}

}
\keyword{internal}
