% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{can_parse_simple}
\alias{can_parse_simple}
\title{Check if an expression can be parsed as simple}
\usage{
can_parse_simple(expression)
}
\arguments{
\item{expression}{Character string with a filter expression}
}
\value{
Logical indicating if expression is simple
}
\description{
Simple expressions are those that can be shown in the simple UI:
\itemize{
\item column == value
\item column != value
\item column > value
\item column < value
\item column >= value
\item column <= value
\item column \%in\% c(...)
\item between(column, min, max)
}
}
\keyword{internal}
