% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs_ui_minimal.R
\name{exprs_ui_minimal}
\alias{exprs_ui_minimal}
\title{Create a minimal UI element for a single expression}
\usage{
exprs_ui_minimal(id = "", value = "")
}
\arguments{
\item{id}{Character string, an identifier for the UI element}

\item{value}{Default value for the expression}
}
\value{
A div element containing the UI components
}
\description{
Create a minimal UI element for a single expression
}
\examples{
\dontrun{
pkgload::load_all(); run_minimal_example()
}
}
