% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{format_range}
\alias{format_range}
\title{Format numeric range filter}
\usage{
format_range(column, range_val, col_range)
}
\arguments{
\item{column}{Column name}

\item{range_val}{Range values}

\item{col_range}{Full column range}
}
\value{
Filter expression string
}
\description{
Format numeric range filter
}
\keyword{internal}
