% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{format_values}
\alias{format_values}
\title{Format value-based filter}
\usage{
format_values(column, values, include_mode = "include")
}
\arguments{
\item{column}{Column name}

\item{values}{Selected values}

\item{include_mode}{"include" or "exclude"}
}
\value{
Filter expression string
}
\description{
Format value-based filter
}
\keyword{internal}
