% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_keyvalue.R
\name{get_exprs}
\alias{get_exprs}
\title{Extract expressions from input values}
\usage{
get_exprs(prefix, input)
}
\arguments{
\item{prefix}{Prefix to match at start of input names}

\item{input}{Shiny input object}
}
\value{
Named character vector of expressions
}
\description{
Extract expressions from input values
}
