% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvexpr.R
\name{get_input_names}
\alias{get_input_names}
\title{Helper function to get input names matching a pattern}
\usage{
get_input_names(prefix, input, regex)
}
\arguments{
\item{prefix}{Prefix to match at start of input names}

\item{input}{Shiny input object}

\item{regex}{Regular expression to match in input names}
}
\value{
Character vector of matching input names
}
\description{
Helper function to get input names matching a pattern
}
