% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_keyvalue.R
\name{get_rms}
\alias{get_rms}
\title{Get remove button values}
\usage{
get_rms(prefix, input)
}
\arguments{
\item{prefix}{Prefix to match at start of input names}

\item{input}{Shiny input object}
}
\value{
Named integer vector of remove button values
}
\description{
Get remove button values
}
