% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_filter.R
\name{\%||\%}
\alias{\%||\%}
\title{Default value operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Default value operator
}
