% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_selector.R
\name{mod_by_selector_server}
\alias{mod_by_selector_server}
\title{Group By Column Selector Server Module}
\usage{
mod_by_selector_server(id, get_cols, initial_value = character())
}
\arguments{
\item{id}{Character string. Module ID.}

\item{get_cols}{Reactive function that returns available column names.}

\item{initial_value}{Character vector. Initial selected columns.}
}
\description{
Group By Column Selector Server Module
}
