% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_enhanced_filter.R
\name{mod_enhanced_filter_ui}
\alias{mod_enhanced_filter_ui}
\title{Create enhanced filter UI module}
\usage{
mod_enhanced_filter_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create enhanced filter UI module
}
\keyword{internal}
