% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpr.R
\name{mod_flexpr_server}
\alias{mod_flexpr_server}
\title{Flexible expression server module}
\usage{
mod_flexpr_server(id, get_value, get_choices, get_choices_expr = get_choices)
}
\arguments{
\item{id}{The module ID}

\item{get_value}{Function that returns initial values}

\item{get_choices}{Function that returns column names for autocompletion}

\item{get_choices_expr}{Function that returns column names for expression autocompletion (defaults to get_choices)}
}
\value{
A reactive expression containing the current expressions
}
\description{
A Shiny module that manages flexible expressions.
}
\examples{
\dontrun{
serve(new_flexpr_block(), list(data = mtcars))
}
}
