% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpr.R
\name{mod_flexpr_ui}
\alias{mod_flexpr_ui}
\title{Create flexible expression UI module}
\usage{
mod_flexpr_ui(
  value,
  cols = c("Sepal.Length", "Sepal.Width"),
  submit = TRUE,
  ns = function(x) x
)
}
\arguments{
\item{value}{Initial values}

\item{cols}{Available column names}

\item{submit}{Whether to show submit button}

\item{ns}{Namespace function}
}
\value{
A div containing the UI elements
}
\description{
Create flexible expression UI module
}
