% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_join_keys.R
\name{mod_join_keys_server}
\alias{mod_join_keys_server}
\title{Join Keys Server Module}
\usage{
mod_join_keys_server(id, get_x_cols, get_y_cols, initial_keys = character())
}
\arguments{
\item{id}{Character string. Module ID.}

\item{get_x_cols}{Reactive function that returns column names from the first dataset.}

\item{get_y_cols}{Reactive function that returns column names from the second dataset.}

\item{initial_keys}{List or character vector. Initial join keys.}
}
\description{
Join Keys Server Module
}
