% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvalue.R
\name{mod_keyvalue_server}
\alias{mod_keyvalue_server}
\title{Key-value server module}
\usage{
mod_keyvalue_server(id, submit = TRUE, multiple = TRUE, key = "suggest")
}
\arguments{
\item{id}{The module ID}

\item{submit}{Whether to show a submit button (defaults to TRUE)}

\item{multiple}{Whether multiple key-value pairs are allowed (defaults to TRUE)}

\item{key}{The key to use (defaults to "suggest")}
}
\description{
Key-value server module
}
