% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_keyvalue.R
\name{mod_keyvalue_ui}
\alias{mod_keyvalue_ui}
\title{Create key-value UI module}
\usage{
mod_keyvalue_ui(
  value,
  multiple,
  submit,
  key,
  auto_complete_list = NULL,
  ns = function(x) x
)
}
\arguments{
\item{value}{Initial values}

\item{multiple}{Whether multiple key-value pairs are allowed}

\item{submit}{Whether to show submit button}

\item{key}{Key display mode}

\item{auto_complete_list}{List of autocomplete options}

\item{ns}{Namespace function}
}
\value{
A div containing the UI elements
}
\description{
Create key-value UI module
}
