% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvexpr.R
\name{mod_kvexpr_ui}
\alias{mod_kvexpr_ui}
\title{Create key-value UI module}
\usage{
mod_kvexpr_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create key-value UI module
}
