% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_arrange.R
\name{mod_multi_arrange_ui}
\alias{mod_multi_arrange_ui}
\title{Create multi arrange UI module}
\usage{
mod_multi_arrange_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create multi arrange UI module
}
