% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_filter.R
\name{mod_multi_filter_ui}
\alias{mod_multi_filter_ui}
\title{Create multi filter UI module}
\usage{
mod_multi_filter_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create multi filter UI module
}
\keyword{internal}
