% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_kvexpr.R
\name{mod_multi_kvexpr_ui}
\alias{mod_multi_kvexpr_ui}
\title{Create multi key-value UI module}
\usage{
mod_multi_kvexpr_ui(id, extra_button = NULL)
}
\arguments{
\item{id}{The module ID}

\item{extra_button}{Optional UI element (e.g., submit button) to display on the right side}
}
\value{
A div containing the UI elements
}
\description{
Create multi key-value UI module
}
\keyword{internal}
