% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_select.R
\name{mod_multi_select_ui}
\alias{mod_multi_select_ui}
\title{Create multi select UI module}
\usage{
mod_multi_select_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create multi select UI module
}
\keyword{internal}
