% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table_select.R
\name{mod_table_select_ui}
\alias{mod_table_select_ui}
\title{Create table select UI module}
\usage{
mod_table_select_ui(id, show_selected_on_top = TRUE)
}
\arguments{
\item{id}{The module ID}

\item{show_selected_on_top}{Logical, whether to show selected columns section on top}
}
\value{
A div containing the UI elements
}
\description{
Create table select UI module
}
