% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_filter.R
\name{mod_value_filter_ui}
\alias{mod_value_filter_ui}
\title{Create value filter UI module}
\usage{
mod_value_filter_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create value filter UI module
}
\keyword{internal}
