% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vexpr.R
\name{mod_vexpr_ui}
\alias{mod_vexpr_ui}
\title{Create value expression UI module}
\usage{
mod_vexpr_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create value expression UI module
}
