% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_filter.R
\name{multi_filter_condition_ui}
\alias{multi_filter_condition_ui}
\title{Create UI for a single filter condition row}
\usage{
multi_filter_condition_ui(id, value = "TRUE", show_remove = TRUE)
}
\arguments{
\item{id}{Row identifier}

\item{value}{Condition value}

\item{show_remove}{Whether to show remove button}
}
\value{
A div containing the row UI
}
\description{
Create UI for a single filter condition row
}
