% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_kvexpr.R
\name{multi_kvexpr_row_ui}
\alias{multi_kvexpr_row_ui}
\title{Create UI for a single expression row}
\usage{
multi_kvexpr_row_ui(id, name = "new_col", value = "1", show_remove = TRUE)
}
\arguments{
\item{id}{Row identifier}

\item{name}{Column name}

\item{value}{Expression value}

\item{show_remove}{Whether to show remove button}
}
\value{
A div containing the row UI
}
\description{
Create UI for a single expression row
}
