% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask.R
\name{new_ask_block}
\alias{new_ask_block}
\title{Ask block constructor}
\usage{
new_ask_block(prompt = character(), ...)
}
\arguments{
\item{prompt}{prompt}
}
\description{
This block allows for data transformations using LLMs
}
