% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_selector.R
\name{null-coalescing}
\alias{null-coalescing}
\alias{\%||\%}
\title{NULL-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side value}

\item{y}{Right-hand side value}
}
\value{
x if x is not NULL, otherwise y
}
\description{
Returns the right-hand side if the left-hand side is NULL, otherwise returns the left-hand side.
}
\keyword{internal}
