% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{parse_filter_string}
\alias{parse_filter_string}
\title{Parse a filter string into condition objects}
\usage{
parse_filter_string(filter_string)
}
\arguments{
\item{filter_string}{Character string with filter conditions}
}
\value{
List of filter_condition objects
}
\description{
Parse a filter string into condition objects
}
\keyword{internal}
