% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpr.R
\name{run_ace_example}
\alias{run_ace_example}
\title{Run example app demonstrating custom autocompletion}
\usage{
run_ace_example()
}
\description{
This function launches a minimal Shiny app that demonstrates the custom
autocompletion functionality with a simple example.
}
\examples{
\dontrun{
pkgload::load_all(); run_ace_example()
}
}
