% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpr.R
\name{run_flexpr_example}
\alias{run_flexpr_example}
\title{Run example app demonstrating flexible expression functionality}
\usage{
run_flexpr_example()
}
\description{
This function launches a Shiny app that demonstrates the flexible expression
module functionality with a simple example.
}
\examples{
\dontrun{
pkgload::load_all(); run_flexpr_example()
}
}
