% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvexpr.R
\name{run_kvexpr_example}
\alias{run_kvexpr_example}
\title{Run example app demonstrating key-value functionality}
\usage{
run_kvexpr_example()
}
\description{
This function launches a Shiny app that demonstrates the key-value module
functionality with a simple example.
}
\examples{
\dontrun{
pkgload::load_all();run_kvexpr_example()
}
}
