% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs_ui_minimal.R
\name{run_minimal_example}
\alias{run_minimal_example}
\title{Run minimal example app}
\usage{
run_minimal_example()
}
\description{
This function launches a minimal Shiny app that demonstrates the basic
expression UI functionality.
}
\examples{
\dontrun{
pkgload::load_all(); run_minimal_example()
}
}
