% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_arrange.R
\name{run_multi_arrange_example}
\alias{run_multi_arrange_example}
\title{Run example app demonstrating multi arrange functionality}
\usage{
run_multi_arrange_example()
}
\description{
Run example app demonstrating multi arrange functionality
}
\examples{
\dontrun{
run_multi_arrange_example()
}
}
\keyword{internal}
