% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_filter.R
\name{run_multi_filter_example}
\alias{run_multi_filter_example}
\title{Run example app demonstrating multi filter functionality}
\usage{
run_multi_filter_example()
}
\description{
Run example app demonstrating multi filter functionality
}
\examples{
\dontrun{
run_multi_filter_example()
}
}
