% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rename.R
\name{run_multi_rename_example}
\alias{run_multi_rename_example}
\title{Run example app demonstrating multi rename functionality}
\usage{
run_multi_rename_example()
}
\description{
Run example app demonstrating multi rename functionality
}
\examples{
\dontrun{
run_multi_rename_example()
}
}
\keyword{internal}
