% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_select.R
\name{run_multi_select_example}
\alias{run_multi_select_example}
\title{Run example app demonstrating multi select functionality}
\usage{
run_multi_select_example()
}
\description{
Run example app demonstrating multi select functionality
}
\examples{
\dontrun{
run_multi_select_example()
}
}
\keyword{internal}
