% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_filter.R
\name{run_value_filter_example}
\alias{run_value_filter_example}
\title{Run example app demonstrating value filter functionality}
\usage{
run_value_filter_example()
}
\description{
Run example app demonstrating value filter functionality
}
\examples{
\dontrun{
run_value_filter_example()
}
}
\keyword{internal}
