% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vexpr.R
\name{run_vexpr_example}
\alias{run_vexpr_example}
\title{Run example app demonstrating value expression functionality}
\usage{
run_vexpr_example()
}
\description{
This function launches a Shiny app that demonstrates the value expression module
functionality with a simple example.
}
\examples{
\dontrun{
pkgload::load_all();run_vexpr_example()
}
}
