% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{split_filter_conditions}
\alias{split_filter_conditions}
\title{Split a compound filter expression into individual conditions}
\usage{
split_filter_conditions(filter_string)
}
\arguments{
\item{filter_string}{Character string with filter conditions}
}
\value{
Character vector of individual filter expressions
}
\description{
Takes a filter string like "mpg > 20 & cyl == 4" and splits it into
individual condition expressions
}
\keyword{internal}
