% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.dplyr-package}
\alias{blockr.dplyr}
\alias{blockr.dplyr-package}
\title{blockr.dplyr: Data Wrangling Blocks for 'blockr.core'}
\description{
Provides access to 'dplyr'-powered data wrangling as blocks for 'blockr.core'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bristolmyerssquibb.github.io/blockr.dplyr/}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Nicolas Bennett \email{nicolas@cynkra.com}
  \item David Granjon \email{david@cynkra.com}
  \item Mike Page \email{mike@cynkra.com}
}

}
\keyword{internal}
