% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockr-input-dep.R
\name{blockr_input_dep}
\alias{blockr_input_dep}
\title{HTML dependency for blockr-input component (code autocomplete)}
\usage{
blockr_input_dep()
}
\value{
An \code{htmltools::tagList} of \code{htmlDependency} objects.
}
\description{
Exported for reuse by other blockr packages that embed the shared
expression-input component.
}
