% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_by_selector.R, R/mod_join_keys.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper function for NULL default}
\usage{
x \%||\% y

x \%||\% y
}
\description{
Helper function for NULL default

Helper function for NULL default
}
\keyword{internal}
