% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexpr.R
\name{mod_flexpr_ui}
\alias{mod_flexpr_ui}
\title{Create flexible expression UI module}
\usage{
mod_flexpr_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create flexible expression UI module
}
