% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_kvexpr.R
\name{mod_multi_kvexpr_ui}
\alias{mod_multi_kvexpr_ui}
\title{Create multi key-value UI module}
\usage{
mod_multi_kvexpr_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create multi key-value UI module
}
