% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_rename.R
\name{mod_multi_rename_ui}
\alias{mod_multi_rename_ui}
\title{Create multi rename UI module}
\usage{
mod_multi_rename_ui(id)
}
\arguments{
\item{id}{The module ID}
}
\value{
A div containing the UI elements
}
\description{
Create multi rename UI module
}
