% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{new_arrange_block}
\alias{new_arrange_block}
\title{Arrange block constructor}
\usage{
new_arrange_block(cols = character(), desc = logical(), ...)
}
\arguments{
\item{cols}{Columns to arrange by}

\item{desc}{Logical vector indicating which columns should be sorted in descending order}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows for arranging data by columns (see
\code{\link[dplyr:arrange]{dplyr::arrange()}}).
}
