% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{new_arrange_block}
\alias{new_arrange_block}
\title{Arrange block constructor}
\usage{
new_arrange_block(columns = character(), desc = "False", ...)
}
\arguments{
\item{columns}{Columns to arrange by.}

\item{desc}{Should columns be sorted in descending order?}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows allows you to order the rows of a data frame by the values
of selected columns (see \code{\link[dplyr:arrange]{dplyr::arrange()}}).
}
