% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{new_filter_block}
\alias{new_filter_block}
\title{Filter block constructor}
\usage{
new_filter_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows for filtering data using flexfilter's interface (see
\code{\link[flexfilter:flexfilter]{flexfilter::flexfilter_server()}}).
}
