% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{new_join_block}
\alias{new_join_block}
\title{Select block constructor}
\usage{
new_join_block(type = character(), by = character(), ...)
}
\arguments{
\item{type}{Join type}

\item{by}{Column(s) tp join on}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows for joining of two \code{data.frame} objects (see
\code{\link[dplyr:mutate-joins]{dplyr::left_join()}}).
}
\examples{
\dontrun{
serve(
  new_board(
    blocks = c(
      c = new_join_block(),
      a = new_dataset_block(),
      b = new_dataset_block()
    ),
    links = links(
      from = c("a", "b"),
      to = c("c", "c"),
      input = c("x", "y")
    )
  )
)
}

}
