% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{new_mutate_block}
\alias{new_mutate_block}
\title{Mutate block constructor}
\usage{
new_mutate_block(strings = character(), ...)
}
\arguments{
\item{strings}{Character vector of expressions}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to add new variables and preserve existing ones
(see \code{\link[dplyr:mutate]{dplyr::mutate()}}).
}
