% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{new_mutate_block}
\alias{new_mutate_block}
\title{Mutate block constructor}
\usage{
new_mutate_block(r_strings, ...)
}
\arguments{
\item{r_strings}{Reactive expression returning character vector of
expressions}

\item{...}{Additional arguments forwarded to \code{\link[=new_block]{new_block()}}}
}
\value{
A block object for mutate operations
}
\description{
This block allows to add new variables and preserve existing ones
(see \code{\link[dplyr:mutate]{dplyr::mutate()}}).
}
\examples{
\dontrun{
serve(new_mutate_block(), list(data = mtcars))
}
}
\seealso{
\code{\link[=new_transform_block]{new_transform_block()}}
}
