% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{new_select_block}
\alias{new_select_block}
\title{Select block constructor}
\usage{
new_select_block(
  columns = character(),
  interface = "table",
  show_selected_on_top = TRUE,
  ...
)
}
\arguments{
\item{columns}{Selected columns}

\item{interface}{Interface type: "table" or "cards" (default "table")}

\item{show_selected_on_top}{Whether to show selected columns on top (default TRUE for table interface)}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to perform column subsetting on \code{data.frame} objects (see
\code{\link[dplyr:select]{dplyr::select()}}).
}
