% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{new_select_block}
\alias{new_select_block}
\title{Select block constructor}
\usage{
new_select_block(columns = character(), ...)
}
\arguments{
\item{columns}{Selected columns}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to perform column subsetting on \code{data.frame} objects (see
\code{\link[dplyr:select]{dplyr::select()}}).
}
