% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{new_summarize_block}
\alias{new_summarize_block}
\title{Summarize block constructor}
\usage{
new_summarize_block(
  group_cols = character(),
  summary_cols = character(),
  stats = character(),
  ...
)
}
\arguments{
\item{group_cols}{Columns to group by}

\item{summary_cols}{Numeric columns to summarize}

\item{stats}{Summary statistics to calculate}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows for grouping data and calculating summary statistics (see
\code{\link[dplyr:summarise]{dplyr::summarise()}}).
}
