% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{parse_value_filter}
\alias{parse_value_filter}
\title{Parse value filter conditions into dplyr expression}
\usage{
parse_value_filter(conditions = list(), preserve_order = FALSE)
}
\arguments{
\item{conditions}{List of filter conditions}

\item{preserve_order}{Logical. If TRUE, preserves order of selected values}
}
\value{
A parsed expression object
}
\description{
Parse value filter conditions into dplyr expression
}
