% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_filter.R
\name{parse_value_filter}
\alias{parse_value_filter}
\title{Parse value filter conditions into dplyr expression}
\usage{
parse_value_filter(conditions = list(), logic_operators = character())
}
\arguments{
\item{conditions}{List of filter conditions}

\item{logic_operators}{Character vector of logic operators ("&" or "|")}
}
\value{
A parsed expression object
}
\description{
Parse value filter conditions into dplyr expression
}
